convert_md: less diversity, better inference
One of the biggest compile costs in Franklin is `convert_md`.
This standarizes the argument types to reduce the number of times
it needs to be compiled (xref https://timholy.github.io/SnoopCompile.jl/stable/snoopi_deep_analysis/#Argument-standardization)
and improves the inferrability for `lxdefs` and a couple of
other internal variables.